+++ src/enum.ts
@@ -4,3 +4,4 @@ export const startServerID = "fakeResponse.startServer";
 export const stopServerID = "fakeResponse.stopServer";
 export const getRoutesListID = "fakeResponse.getRoutesList";
 export const switchEnvironmentID = "fakeResponse.switchEnvironment";
+export const sortJsonID = "fakeResponse.sortJson";